Debugging enhanced
Definition of DEBUG in the main sketch (easy to change)
DBG definition
Use Serial2 instead of Serial1 for the ESP32
Changed the concept of UART.IF in the -5 sketch: Debug is always Serial and the ESP32 gets an additional Serial2 for communication

esp_now_peer_info_t peerInfo; seems not needed anymore

added a ROLE definition. With this definition it should be possible to have the same directory for all roles. We can discuss this in Skype if you want