Update .gitignore
Added fdrs_globals.h to .gitignore.
This way one can directly work in the repo and maintain personal configuration without exposing it to the public by accidentially submitting them.
If someone knows a safe way to revert a file in a repo with fully removing the previous commit, I would be glad to hear about a solutin.