Fixed issue with ignored global settings in fdrs_sensor.h
This needs to be checked, especially in the new class-based model of sensors and gateways: First the global settings must be checked and the corresponding local settings for a sensor / gateway commented out.
Then either the one or other is saved in a FDRS_prefix variable. After that checks must soley be done against FDRS_variables.