Fixing configuration system - part 1: sensors
Fixed the configuration system for sensor nodes. Gateway nodes are still to be fixed (work still in process :) )
This has been addressed in #71

Important: For this system to work and make the onboarding easier for new users, global settings must be used by default and therefore local settings must be commented out by default.